# api/openapi.json
{
 "openapi": "3.0.3",
 "info": {
 "title": "xrelease API",
 "description": "Self-hosted release notifier — management API and inbound webhooks.\n\nStarted with `xrelease serve`. Polling continues in the background; webhooks provide low-latency push for GitHub, GitLab, Gitea, Codeberg, Bitbucket, and Docker Hub.\n\nConfiguration is split: `bootstrap.toml` (infra) + `app/releases.yaml` (sources/teams/notifiers). Source kinds (`type` in app config): `github`, `codeberg`, `gitea`, `gitlab`, `bitbucket`, `docker`, `ghcr`, `quay`, `ecr`, `feed`, `pypi`, `npm`, `yarn`, `cargo`, `maven`, `nuget`, `hex`, `rubygems`, `packagist`, `cpan`, `artifacthub`.",
 "version": "0.1.0",
 "license": {
 "name": "Apache-2.0",
 "url": "https://www.apache.org/licenses/LICENSE-2.0"
 }
 },
 "servers": [
 {
 "url": "http://127.0.0.1:8080",
 "description": "Default local bind (`[api].listen`)"
 }
 ],
...
 "paths": {
...
 "/api/v1/config": {
 "get": {
 "tags": [
 "observability"
 ],
 "summary": "Effective running config (secrets redacted)",
 "operationId": "getConfig",
 "security": [
 {
 "bearerAuth": []
 }
 ],
 "responses": {
 "200": {
 "description": "Merged bootstrap + desired state with provenance",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ConfigView"
 }
 }
 },
 "headers": {
 "ETag": {
 "description": "SHA-256 of the running desired document; send back as If-Match on apply",
 "schema": {
 "type": "string"
 }
 }
 }
 },
 "401": {
 "$ref": "#/components/responses/Unauthorized"
 }
 }
 }
 },
...
 "/api/v1/organizations": {
 "get": {
 "tags": [
 "organizations"
 ],
 "summary": "Organization catalogue ([[organizations]] in bootstrap.toml)",
 "operationId": "listOrganizations",
 "security": [
 {
 "bearerAuth": []
 }
 ],
 "responses": {
 "200": {
 "description": "Catalogue with live per-org source counts",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/OrganizationListResponse"
 }
 }
 }
 },
 "401": {
 "description": "Missing/invalid bearer token",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 }
 }
 }
 },
 "/api/v1/organizations/{id}/config": {
 "get": {
 "tags": [
 "organizations"
 ],
 "summary": "One organization's desired document from its authority (redacted)",
 "operationId": "getOrganizationConfig",
 "security": [
 {
 "bearerAuth": []
 }
 ],
 "parameters": [
 {
 "name": "id",
 "in": "path",
 "required": true,
 "schema": {
 "type": "string"
 },
 "description": "Organization id ([[organizations]].id)."
 }
 ],
 "responses": {
 "200": {
 "description": "Desired document; ETag header carries content_sha256 for If-Match",
 "headers": {
 "ETag": {
 "schema": {
 "type": "string"
 },
 "description": "Quoted content_sha256 of the raw document."
 }
 },
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/OrganizationConfigView"
 }
 }
 }
 },
 "404": {
 "description": "Unknown organization (or single-document instance)",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 }
 }
 }
 },
 "/api/v1/organizations/{id}/config/revisions": {
 "get": {
 "tags": [
 "organizations"
 ],
 "summary": "This organization's ledger stream (metadata only)",
 "operationId": "listOrganizationRevisions",
 "security": [
 {
 "bearerAuth": []
 }
 ],
 "parameters": [
 {
 "name": "id",
 "in": "path",
 "required": true,
 "schema": {
 "type": "string"
 },
 "description": "Organization id ([[organizations]].id)."
 },
 {
 "name": "limit",
 "in": "query",
 "schema": {
 "type": "integer",
 "minimum": 1,
 "maximum": 200
 }
 },
 {
 "name": "offset",
 "in": "query",
 "schema": {
 "type": "integer",
 "minimum": 0
 }
 }
 ],
 "responses": {
 "200": {
 "description": "Page of this org's apply/reject history",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/OrganizationRevisionsResponse"
 }
 }
 }
 },
 "404": {
 "description": "Unknown organization",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 }
 }
 }
 },
 "/api/v1/organizations/{id}/config/validate": {
 "post": {
 "tags": [
 "organizations"
 ],
 "summary": "Dry-run one organization's candidate document against the composed runtime",
 "operationId": "validateOrganizationConfig",
 "security": [
 {
 "bearerAuth": []
 }
 ],
 "parameters": [
 {
 "name": "id",
 "in": "path",
 "required": true,
 "schema": {
 "type": "string"
 },
 "description": "Organization id ([[organizations]].id)."
 }
 ],
 "requestBody": {
 "required": true,
 "content": {
 "application/yaml": {
 "schema": {
 "type": "string"
 }
 },
 "application/toml": {
 "schema": {
 "type": "string"
 }
 }
 },
 "description": "Complete desired-state document for this organization (YAML or TOML)."
 },
 "responses": {
 "200": {
 "description": "Validation report (valid may be false)",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ConfigValidateResponse"
 }
 }
 }
 },
 "400": {
 "description": "Document does not parse / carries infra sections",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
 "403": {
 "description": "Authenticated but the principal's role is insufficient (RBAC).",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
 "404": {
 "description": "Unknown organization",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 }
 }
 }
 },
 "/api/v1/organizations/{id}/config/apply": {
 "post": {
 "tags": [
 "organizations"
 ],
 "summary": "Replace one organization's desired document (hot-swap + org ledger stream)",
 "description": "Requires [config_api].api_config = true and source = \"api\". If-Match compares against this organization's current authority sha (ETag of GET./config).",
 "operationId": "applyOrganizationConfig",
 "security": [
 {
 "bearerAuth": []
 }
 ],
 "parameters": [
 {
 "name": "id",
 "in": "path",
 "required": true,
 "schema": {
 "type": "string"
 },
 "description": "Organization id ([[organizations]].id)."
 },
 {
 "name": "If-Match",
 "in": "header",
 "required": false,
 "schema": {
 "type": "string"
 },
 "description": "Optimistic concurrency against this org's current document sha."
 },
 {
 "name": "X-Config-Revision",
 "in": "header",
 "required": false,
 "schema": {
 "type": "string"
 },
 "description": "Audit label (e.g. a git SHA)."
 }
 ],
 "requestBody": {
 "required": true,
 "content": {
 "application/yaml": {
 "schema": {
 "type": "string"
 }
 },
 "application/toml": {
 "schema": {
 "type": "string"
 }
 }
 },
 "description": "Complete desired-state document for this organization (YAML or TOML)."
 },
 "responses": {
 "200": {
 "description": "Applied (or idempotent no-op)",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ConfigApplyResponse"
 }
 }
 }
 },
 "400": {
 "description": "Document does not parse / carries forbidden sections",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
 "403": {
 "description": "Authenticated but the principal's role is insufficient (RBAC).",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
 "404": {
 "description": "Unknown organization, or [config_api].api_config is off",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
 "409": {
 "description": "source = \"local\" pins desired state to Git",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
 "412": {
 "description": "If-Match precondition failed",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
 "422": {
 "description": "Document parsed but the composed runtime failed validation",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 }
 }
 }
 },
 "/api/v1/organizations/{id}/config/rollback": {
 "post": {
 "tags": [
 "organizations"
 ],
 "summary": "Re-apply the previous applied revision of this organization's stream",
 "operationId": "rollbackOrganizationConfig",
 "security": [
 {
 "bearerAuth": []
 }
 ],
 "parameters": [
 {
 "name": "id",
 "in": "path",
 "required": true,
 "schema": {
 "type": "string"
 },
 "description": "Organization id ([[organizations]].id)."
 }
 ],
 "responses": {
 "200": {
 "description": "Rolled back",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ConfigApplyResponse"
 }
 }
 }
 },
 "403": {
 "description": "Authenticated but the principal's role is insufficient (RBAC).",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
 "404": {
 "description": "Unknown organization, or [config_api].api_config is off",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
 "409": {
 "description": "No earlier applied revision in this stream, or source = \"local\"",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 }
 }
 }
 }
 },
 "components": {
 "securitySchemes": {
 "bearerAuth": {
 "type": "http",
 "scheme": "bearer",
 "description": "Static API key (`[api].api_key` / `XRELEASE_API_KEY`) or OIDC access token JWT (`[api.oidc]` / `XRELEASE_OIDC_ISSUER`). Required when either auth method is configured."
 }
 },
 "responses": {
 "Unauthorized": {
 "description": "Authentication failed",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
...
 },
 "schemas": {
...
 "ErrorResponse": {
 "type": "object",
 "properties": {
 "error": {
 "type": "string"
 }
 },
 "required": [
 "error"
 ]
 },
...
 "ConfigView": {
 "type": "object",
 "properties": {
 "content": {
 "type": "string",
 "description": "Redacted effective TOML (bootstrap + desired)"
 },
 "desired_content": {
 "type": "string",
 "nullable": true,
 "description": "Redacted desired-state document only"
 },
 "desired_source": {
 "type": "string",
 "enum": [
 "ledger",
 "app_file",
 "empty"
 ]
 },
 "content_sha256": {
 "type": "string",
 "nullable": true
 },
 "revision": {
 "type": "integer",
 "format": "int64",
 "nullable": true
 },
 "revision_label": {
 "type": "string",
 "nullable": true
 },
 "applied_at": {
 "type": "string",
 "nullable": true
 },
 "bootstrap_path": {
 "type": "string"
 },
 "api_config_enabled": {
 "type": "boolean",
 "description": "Whether POST apply/rollback routes are mounted ([config_api].api_config)."
 },
 "bootstrap_sections": {
 "type": "array",
 "items": {
 "type": "string"
 }
 },
 "ui_config_enabled": {
 "type": "boolean",
 "description": "Whether the dashboard should offer a config editor ([config_api].ui_config, which also requires api_config and source=api)."
 },
 "config_source": {
 "type": "string",
 "enum": [
 "local",
 "api"
 ],
 "description": "Declared authority for desired state ([config_api].source). `local` refuses pushed config with 409."
 }
 },
 "required": [
 "content",
 "desired_source",
 "bootstrap_path",
 "api_config_enabled",
 "bootstrap_sections"
 ]
 },
 "ConfigValidateResponse": {
 "type": "object",
 "properties": {
 "valid": {
 "type": "boolean"
 },
 "content_sha256": {
 "type": "string"
 },
 "report": {
 "type": "object"
 }
 },
 "required": [
 "valid",
 "content_sha256",
 "report"
 ]
 },
 "ConfigApplyResponse": {
 "type": "object",
 "properties": {
 "applied": {
 "type": "boolean"
 },
 "content_sha256": {
 "type": "string"
 },
 "revision": {
 "type": "integer",
 "format": "int64"
 },
 "sources_added": {
 "type": "array",
 "items": {
 "type": "string"
 }
 },
 "sources_removed": {
 "type": "array",
 "items": {
 "type": "string"
 }
 },
 "sources_changed": {
 "type": "array",
 "items": {
 "type": "string"
 }
 },
 "warnings": {
 "type": "array",
 "items": {
 "type": "string"
 }
 }
 },
 "required": [
 "applied",
 "content_sha256",
 "revision"
 ]
 },
...
 "ConfigRevisionSummary": {
 "type": "object",
 "description": "One ledger row without its config body.",
 "properties": {
 "id": {
 "type": "integer",
 "format": "int64"
 },
 "content_sha256": {
 "type": "string"
 },
 "revision_label": {
 "type": "string",
 "nullable": true
 },
 "applied_at": {
 "type": "string",
 "format": "date-time"
 },
 "applied_by": {
 "type": "string",
 "nullable": true,
 "description": "Server-verified principal, e.g. `api_key` or `oidc:<sub>`"
 },
 "source_addr": {
 "type": "string",
 "nullable": true
 },
 "status": {
 "type": "string",
 "enum": [
 "applied",
 "rejected"
 ]
 },
 "error": {
 "type": "string",
 "nullable": true
 },
 "organization_id": {
 "type": "string",
 "nullable": true,
 "description": "Ledger stream: absent/null = legacy single-document stream, else the [[organizations]] id."
 }
 },
 "required": [
 "id",
 "content_sha256",
 "applied_at",
 "status"
 ]
 },
...
 "OrganizationView": {
 "type": "object",
 "required": [
 "id",
 "name",
 "source_count"
 ],
 "properties": {
 "id": {
 "type": "string",
 "description": "Catalogue slug ([[organizations]].id)."
 },
 "name": {
 "type": "string",
 "description": "Display name (falls back to id)."
 },
 "source_count": {
 "type": "integer",
 "description": "Watches currently running for this organization."
 }
 }
 },
 "OrganizationListResponse": {
 "type": "object",
 "required": [
 "organizations",
 "config_source",
 "api_config_enabled"
 ],
 "properties": {
 "organizations": {
 "type": "array",
 "items": {
 "$ref": "#/components/schemas/OrganizationView"
 }
 },
 "config_source": {
 "type": "string",
 "enum": [
 "local",
 "api"
 ],
 "description": "Desired-state authority declared in [config_api]."
 },
 "api_config_enabled": {
 "type": "boolean"
 },
 "ui_config_enabled": {
 "type": "boolean",
 "description": "Whether the dashboard form editor is on (ui_config + api_config + source=api)."
 }
 }
 },
 "OrganizationConfigView": {
 "type": "object",
 "required": [
 "organization",
 "name",
 "desired_source",
 "content_sha256"
 ],
 "properties": {
 "organization": {
 "type": "string"
 },
 "name": {
 "type": "string"
 },
 "desired_source": {
 "type": "string",
 "enum": [
 "ledger",
 "app_file",
 "empty"
 ],
 "description": "Authority the document was read from."
 },
 "content_sha256": {
 "type": "string",
 "description": "SHA-256 of the raw desired document; also sent as the ETag header for If-Match."
 },
 "desired_content": {
 "type": "string",
 "nullable": true,
 "description": "Redacted desired document."
 },
 "app_path": {
 "type": "string",
 "nullable": true,
 "description": "[[organizations]].app path (seed / GitOps authority). Null when omitted for API/UI-authored orgs."
 },
 "config_source": {
 "type": "string",
 "enum": [
 "local",
 "api"
 ]
 },
 "accepts_pushed_config": {
 "type": "boolean"
 },
 "api_config_enabled": {
 "type": "boolean",
 "description": "Whether POST apply/rollback routes are mounted ([config_api].api_config)."
 },
 "ui_config_enabled": {
 "type": "boolean",
 "description": "Whether the dashboard should offer the form editor ([config_api].ui_config)."
 }
 }
 },
 "OrganizationRevisionsResponse": {
 "type": "object",
 "required": [
 "organization",
 "revisions",
 "total",
 "limit",
 "offset"
 ],
 "properties": {
 "organization": {
 "type": "string"
 },
 "revisions": {
 "type": "array",
 "items": {
 "$ref": "#/components/schemas/ConfigRevisionSummary"
 }
 },
 "total": {
 "type": "integer"
 },
 "limit": {
 "type": "integer"
 },
 "offset": {
 "type": "integer"
 }
 }
 },
...
 }
 }
}